Rename SmallPiecewiseBuilder.add -> add_overlay and return Self
As I have so far decided to make the SmallPiecewiseBuilder
consume itself on build(), I should make the add function compatible
with such use.  Operating on self by reference and returning self
reference does not permit the final build() to consume the Builder.

Second, I renamed add -> add_overlay, in an attempt to clarify
the behavior - in which new ValueOverIntervals (should they
overlap in Domain with existing defined Domain) over-write the
value (overlay).

E.g. If the Builder already contains an Unbounded interval with
value 5.0, and a new add_overlay gives a value of -1.0 for the
closed interval [1, 4] -> then the Builder should contain three
ValueOverIntervals as a result of the operation:
[-inf, 1) @ 5.0
[1, 4] @ -1.0
(4, inf) @ 5.0